extract/requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
google-generativeai==0.3.0
openai==1.3.0
pdfplumber==0.10.3
pandas==2.1.1
requests==2.31.0
python-multipart==0.0.6
